third_party/WebKit/LayoutTests/css3/device-adapt/opera/cssom-001-expected.txt


FAIL CSS Test: @viewport CSSOM - CSSRule.VIEWPORT_RULE assert_equals: expected 15 but got undefined
